+++ dogfooding/lib/di/injector.dart
@@ -179,9 +179,6 @@ StreamVideo _initStreamVideo(
       ),
       pushConfiguration: const StreamVideoPushConfiguration(
         ios: IOSPushConfiguration(iconName: 'IconMask'),
-        android: AndroidPushConfiguration(
-          defaultAvatar: 'assets/logo.png',
-        ),
       ),
       registerApnDeviceToken: true,
     ),