+++ packages/stream_video_push_notification/lib/src/stream_video_push_provider.dart
@@ -75,7 +75,7 @@ final class StreamTokenProvider {
   static Stream<String> get onVoIPTokenRefresh {
     if (!CurrentPlatform.isIos) return const Stream.empty();
 
-    return StreamCallKit()
+    return RingingEventBroadcaster()
         .onEvent
         .whereType<ActionDidUpdateDevicePushTokenVoip>()
         .map((event) => event.token);