+++ packages/stream_video_push_notification/lib/stream_video_push_notification.dart
@@ -1,6 +1,6 @@
 library stream_video_push_notification;
 
 export 'src/stream_video_push_notification.dart'
-    hide StreamTokenProvider, StreamCallKit;
+    hide StreamTokenProvider, RingingEventBroadcaster;
 export 'src/stream_video_push_params.dart';
 export 'src/stream_video_push_configuration.dart';